parse: case constants must be int64_t, not int
`int begin = const_expr_val(...)` truncated 64-bit case values to 32 bits.
For values with bit 31 set (e.g. ioctl numbers like 0x80047410UL), the
truncation produced a negative int that sign-extended on assignment to
node->begin (long) — yielding 0xFFFFFFFF80047410. codegen_arm64 then
emitted a 4-movk sequence with the upper 32 bits set, and the runtime
cmp against an unsigned switch expression never matched. Every such
case fell through to default.

Surfaced as: NetBSD/aarch64 kernel built with ncc returned ENOSYS for
every tty ioctl whose number had the IOW/IOR direction bit set, because
ttioctl()'s big switch in kern/tty.c stopped matching those cases —
getty couldn't open /dev/console, no login prompt. With this fix the
kern/tty.c gcc workaround is no longer needed.

Regression test in tests/regression/18_switch_case_high_bit_const.c.
30/30 compliance suite passes; stage1 == stage2 bootstrap fixed point
holds.

+++ src/parse.c
@@ -4160,9 +4160,12 @@ static Node *stmt(Token **rest, Token *tok) {
       error_tok(tok, "stray case");
 
     Node *node = new_node(ND_CASE, tok);
-    int begin = const_expr_val(&tok, tok->next);
+    // Use int64_t — truncating to int sign-extends 32-bit values with bit 31
+    // set when assigned to node->begin (which is long), making case
+    // constants like ioctl numbers (0x80047410UL) compare unequal.
+    int64_t begin = const_expr_val(&tok, tok->next);
 
-    int end;
+    int64_t end;
     if (equal(tok, "...")) {
       // GCC case range extension: case 1 ... 5
       end = const_expr_val(&tok, tok->next);
